TEST: Throw invalid weapon exception when trying to equip wrong level weapon

+++ rpg-characters/Helpers/WeaponAttributes.cs
@@ -5,6 +5,6 @@ namespace rpg_characters.Helpers
     public class WeaponAttributes
     {
         public int Damage { get; set; }
-        public int AttackSpeed { get; set; }
+        public double AttackSpeed { get; set; }
     }
 }
